ta: Compile TA with "-fno-stack-protector" if TA stack protector is off
Depends on compiler version and OS distributions, the stack protector
may be enabled by default. This means TA is always compiled with
"-fstack-protector", even if _CFG_TA_STACK_PROTECTOR is not 'y'.

To solve this issue, we explicitly provide "-fno-stack-protector"
compile option when _CFG_TA_STACK_PROTECTOR is not 'y'. Thus, the TA
stack protector is definitely disabled.

Signed-off-by: Alvin Chang <alvinga@andestech.com>
Reviewed-by: Yu-Chien Peter Lin <peter.lin@sifive.com>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>