libutils: merge malloc pools if possible
When adding pools with for instance malloc_add_pool(), check if the pool
can be merged with other pools. This decreases fragmentation by removing
unneeded allocated end of pool sentinels between merged pools.

The end of pool sentinel is changed from a struct bhead to a struct
bfhead, increasing it by two pointers, to allow it to be freed using
brel() once it has been turned into a regular allocated block.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>