keys: increase default RSA key size to 4096 bits
Change the key size in your default key located at default.pem to 4096
bits.

New key has been created using this command:
  openssl genrsa -out keys/default.pem 4096

Background:
GlobalPlatform, based on feedback from various national bodies such as
ANSSI, BSI, SOGIS, and NIST, has decided to designate RSA keys smaller
than 2048 bits as deprecated (see [1]).

Note:
This key is intended for testing purposes only. Therefore, it's not a
problem to publicly publish this, but it's important to remember to
change this if/when used in real products.

Link: https://globalplatform.org/specs-library/globalplatform-technology-cryptographic-algorithm-recommendations/ [1]

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>