plat-k3: drivers: Protect OP-TEE memory from normal world
This protects the OP-TEE memory from accesses from the normal world.
Without this patch the whole OP-TEE memory is accessible both read and
write from the normal world. This patch has been tested on AM62x. The
other K3 platforms likely suffer from the same problem, but I don't
have a way to test any of them, so these are untouched for now.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>