core: add boot_mem section for pager
With CFG_WITH_PAGER=y, a special section of memory is carved out by the
linker script to be used by the boot_mem*() functions during boot. The
amount of memory must cover the worst case, but the remaining unused
memory will be returned to be managed by the pager at end of boot in a
later patch. This enables boot configuration in the same way regardless
of CFG_DYN_CONFIG.

Add support in boot_mem_release_unused() and
boot_mem_release_tmp_alloc() to pass the unused memory to the pager.
This is needed in later patches with CFG_WITH_PAGER=y and
CFG_DYN_CONFIG=y.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>