drivers: stm32_tamp: add tamper events detection support
The anti-tamper detection circuit is used to protect sensitive data
from external attacks. The backup registers, as well as other secrets in
the device, are protected by this anti-tamper detection circuit with
some tamper pins and internal tampers. The external tamper pins can
be configured for edge detection, or level detection with or without
filtering, or active tamper which increases the security level by auto
checking that the tamper pins are not externally opened or shorted.

Add support for the device-tree configuration of internal and external
tamper events as well as the list hardware mapped pins with their
associated tamper ID for external tampers.

While there, fix the license for this driver.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>