dts: stm32: add tamper event detection configuration for stm32mp135f-dk
Add and default enable support for the TAMP button present on the
stm32mp135f-dk board. It relies on the external tamper 2.
Set GPIOA6 as secure as it now serve this purpose.

Add and default disable support for a test setup of an active tamper
event detection that is feasible with the GPIO expansion present on the
stm32mp135f-dk board.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>