.gitignore: remove specific IDE configurations directory
Every developer has its own configuration directory, vscode is not an
exception, so remove `.vscode` configuration directory from the
gitignore exception list.
VSCode's `extensions.json` file is not compulsory for OP-TEE
development.
Remove the file as it is used for an extension recommendation pop up
inside VSCode, and it can mess with people's VSCode own settings.

Signed-off-by: Clément Le Goffic <clement.legoffic@foss.st.com>