[RFC] ta: pkcs11: Panic on init failure
Panic on any failure to initialize the PKCS#11 TA. The initialization is
only run once (regardless of failures). Hence, we need to panic to avoid
future clients invoking functions on an uninitialized TA instance.

The initialization of a TA seems to only run once, regardless if it
failed or not:

```c
static TEE_Result ta_header_add_session(uint32_t session_id) {
---8<---
	if (!init_done) {
		init_done = true;
		res = init_instance();
		if (res)
			return res;
	}
---8<---
```

If `tee-supplicant` is not running, the `init_persistent_db()` function
part of pkcs11 TA's init will fail. This means the token's `db_main` and
`db_objs` are `NULL`. For the first client, this is fine, as it will fail
to open the session. But for the second client, this will cause issues if
it invokes any function that accesses any part of a token that required
setup. This will cause an abort due to bad memory access.

Signed-off-by: Johan Korsnes <johan.korsnes@remarkable.no>