mk/config.mk: allow CFG_CALLOUT for CFG_CORE_SEL2_SPMC=y
Upstream Hafnium provides emulated EL1 physical timer to be used by
OP-TEE. This timer is used by the callout service. However, the current
Hafnium is not stable with this configuration. When creating and later
destroying guest it sometimes crashes with the following output from
Hafnium:
Panic: ASSERT: ../../src/ffa/spmc/interrupts.c:271

So by default keep it disabled, but allow it to be overridden for
testing.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>