core: imx: add generic boards
For i.MX we have to specify the exact board we want to build. Most of
the time a board support merely consists of setting CFG_UART_BASE and
CFG_DDR_SIZE for that board. These variables could be specified from the
outside by the build system as well.

Instead of forcing the user to patch OP-TEE for each new board allow to
build optee generically for a i.MX SoC by adding the SoC name to the
flavorlists. For example an OP-TEE binary for i.MX6UL can then be built
with

make PLATFORM=imx-mx6ul CFG_UART_BASE=UART1_BASE CFG_DDR_SIZE=0x10000000

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>