core: drivers: allow to configure RIF for GPIO not on the SoC package
The number of GPIO physically accessible depends on package.
Some GPIO pins might not be accessible but it is still possible to write
RIF registers to block access.

The assert(nb_rif_conf <= bank->ngpios) mandate to have less or the same
number of RIF configuration than the number of GPIO pin describe with
the property gpio-ranges to have the same number.

Remove the assert and replace() it with MIN() to be less restrictive.

Signed-off-by: Thomas Bourgoin <thomas.bourgoin@foss.st.com>