Fix potential panic in trace / print handler
While porting to a new platform I ended up with some broken mappings
during dom0less NS init and serial_console->ops ended up unmapped.
This meant the DMSG() in do_init_calls() would invariably panic
at that time. This adds a check to avoid trying to call the
unmapped ops vector.

Without this check the result is a recursive panic in the console
output code with no output, which can be tedious to debug on actual
hardware.

If everything is located in the right place and there is sufficient VA
available, this should never happen. So this is really just a quality
of life improvement for someone getting optee to work in a new situation.

Signed-off-by: Jan Setje-Eilers <jan.setjeeilers@oracle.com>