drivers: stm32_rtc: prevent registering of RTC interrupt when it's not used by OP-TEE
Do not register the RTC interrupt when it's not used by OP-TEE as it
prevents non-secure world from handling that interrupt.

The RTC interrupt line is only used if the RTC is set as a wakeup source
or if the RTC is secured by the RIF (only for STM32MP2x)
On STM32MP1, Linux is using the wakeup capabilities of the RTC and
requires the interrupt, so OP-TEE should not register the interrupt line
and securise it.
To prevent OP-TEE from setting the interrupt line as secure when Linux
will need it, we only register the interrupt line when the property
"wakeup-source" is present in the DT or if the RIF configuration the RTC
as secure.

Signed-off-by: Thomas Bourgoin <thomas.bourgoin@foss.st.com>
Reviewed-by: Clément Le Goffic <legoffic.clement@gmail.com>