[WIP] ci: notify reviewers/maintainers for specific subsystems
Introduce a new Python script called notify_maintainers.py and run it in
CI. What it does is simple: when invoked in CI and in the context of a
pull request, it uses get_maintainer.py to find out the GitHub handles
of all the persons who may be concerned by the changes made in the PR
and posts a comment so that these persons may be notified via email.
People are tagged only once per PR (they normally receive subsequent
messages automatically). PR author, assignees and requested reviewers
(if any) are skipped, as well as the default maintainers ("THE REST").

FIXME: People who have their GitHub handle in the MAINTAINERS file and
who are already subscribed to receive PR notifications will most likely
receive two emails upon creation of a PR that touches their area of
expertise -- one when the PR is created, and one shortly after when the
script runs and the GitHub bot tags them in a comment. This need to be
checked.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>