macros.mk: fix version string generation for clang
When the clang supports HIP and detects its installed it will print
its version as part of the `-v` command, which makes the grep return
two matched lines, one of which is unrelated to the clang version.
Fix by only taking the first match, which for clang its always the
version, this should not affect GCC builds as they return a single
match anyways.

Signed-off-by: Itay Almog <itay2828@gmail.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>