ci: notify_maintainers: fix source code comparison
On a pull_request_target event, the code checked out by actions/checkout
is the base branch. Therefore, to compare the PR branch against the base
branch tj-actions/changed-files needs to be given a reference to the PR
head. In other words, base_sha: should actually be the PR head, not the
base. Change it accordingly. Fixes the following error:

 Similar commit hashes detected: previous sha: <xxx> is equivalent to
 the current sha: <xxx>

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>